四則演算に関する不具合あり、整数以外の計算にも対応　8/26　15：27

+++ CALC/calculation.cs
@@ -8,7 +8,7 @@ namespace CALC
 {
     class KEISANNKI
     {
-        public string Super_Easy_calc(int x,int y, char c)
+        public string Super_Easy_calc(double x,double y, char c)
         {
             //計算機プログラム　超簡易型
             if (c == '+')
@@ -54,7 +54,6 @@ namespace CALC
             return return_value;
         }
         //一行で入力可能なプログラム  ()演算子も扱えるようにする
-        //ただしこのプログラムだと答えが整数になる計算しかできない　要修正
         public string Calc(string formula)
         {
             char[] op = new char[] { '(', ')', '*', '/', '+', '-' };//計算記号をまとめておく
@@ -98,16 +97,16 @@ namespace CALC
                         if (exist_op_front == -1)
                         {
                             //前にも記号がないとき　
-                            int x = int.Parse(Splitstring(formula, start, exist_op_times - 1));
-                            int y = int.Parse(Splitstring(formula, exist_op_times + 1, end));
+                            double x = double.Parse(Splitstring(formula, start, exist_op_times - 1));
+                            double y = double.Parse(Splitstring(formula, exist_op_times + 1, end));
                             char c = formula[exist_op_times];
                             return_value = Super_Easy_calc(x, y, c);//前後に記号がなければ計算はそれ以上発生しないので返り値に入れてしまう
                             return return_value;
                         }
                         else {
                             //前には記号があるとき
-                            int x = int.Parse(Splitstring(formula, exist_op_front + 1, exist_op_times - 1));
-                            int y = int.Parse(Splitstring(formula, exist_op_times + 1, end));
+                            double x = double.Parse(Splitstring(formula, exist_op_front + 1, exist_op_times - 1));
+                            double y = double.Parse(Splitstring(formula, exist_op_times + 1, end));
                             char c = formula[exist_op_times];
                             ans = Super_Easy_calc(x, y, c);
                             return_value = Splitstring(formula, start, exist_op_front) + ans;
@@ -120,8 +119,8 @@ namespace CALC
                         if (exist_op_front == -1)
                         {
                             //前には記号がないとき　
-                            int x = int.Parse(Splitstring(formula, start, exist_op_times - 1));
-                            int y = int.Parse(Splitstring(formula, exist_op_times + 1, min_op_pos-1));
+                            double x = double.Parse(Splitstring(formula, start, exist_op_times - 1));
+                            double y = double.Parse(Splitstring(formula, exist_op_times + 1, min_op_pos-1));
                             char c = formula[exist_op_times];
                             ans = Super_Easy_calc(x, y, c);
                             return_value = ans + Splitstring(formula, min_op_pos, end);
@@ -130,8 +129,8 @@ namespace CALC
                         else
                         {
                             //前にも記号があるとき
-                            int x = int.Parse(Splitstring(formula, exist_op_front + 1, exist_op_times - 1));
-                            int y = int.Parse(Splitstring(formula, exist_op_times + 1, min_op_pos - 1));
+                            double x = double.Parse(Splitstring(formula, exist_op_front + 1, exist_op_times - 1));
+                            double y = double.Parse(Splitstring(formula, exist_op_times + 1, min_op_pos - 1));
                             char c = formula[exist_op_times];
                             ans = Super_Easy_calc(x, y, c);
                             return_value = Splitstring(formula, start, exist_op_front) + ans+Splitstring(formula,min_op_pos,end);
@@ -152,13 +151,13 @@ namespace CALC
                     if (exist_op_front != -1)
                     {
                         //前に記号があった時
-                        int x = int.Parse(Splitstring(formula, exist_op_front + 1, exist_op_div - 1));
+                        double x = double.Parse(Splitstring(formula, exist_op_front + 1, exist_op_div - 1));
                         char c = formula[exist_op_div];
                         if (exist_plus_back == -1 && exist_minus_back == -1&&exist_div_back==-1)
                         {
                             //後ろに記号がなかった時
 
-                            int y = int.Parse(Splitstring(formula, exist_op_div + 1, end));
+                            double y = double.Parse(Splitstring(formula, exist_op_div + 1, end));
 
                             ans = Super_Easy_calc(x, y, c);
                             return_value = Splitstring(formula, start, exist_op_div - 1) + ans;
@@ -168,7 +167,7 @@ namespace CALC
                         {
                             //後ろに記号があった時
                             int min_op_pos =Search_pos_min(exist_back) + exist_op_div + 1;
-                            int y = int.Parse(Splitstring(formula, exist_op_div + 1, min_op_pos - 1));
+                            double y = double.Parse(Splitstring(formula, exist_op_div + 1, min_op_pos - 1));
                             ans = Super_Easy_calc(x, y, c);
                             return_value = Splitstring(formula, start, exist_op_div - 1) + ans + Splitstring(formula, min_op_pos, end);
                             return Calc(return_value);
@@ -176,13 +175,13 @@ namespace CALC
                     }
                     else {
                         //前に記号がなかった時
-                        int x = int.Parse(Splitstring(formula,start,exist_op_div-1));
+                        double x = double.Parse(Splitstring(formula,start,exist_op_div-1));
                         char c = formula[exist_op_div];
                         if (exist_plus_back == -1 && exist_minus_back == -1&&exist_div_back==-1)
                         {
                             //後ろに記号がなかった時 最後の計算なので値を返す
 
-                            int y = int.Parse(Splitstring(formula, exist_op_div + 1, end));
+                            double y = double.Parse(Splitstring(formula, exist_op_div + 1, end));
 
                             ans = Super_Easy_calc(x, y, c);
                             return_value =ans;
@@ -192,7 +191,7 @@ namespace CALC
                         {
                             //後ろに記号があった時
                             int min_op_pos = Search_pos_min(exist_back) + exist_op_div + 1;
-                            int y = int.Parse(Splitstring(formula, exist_op_div + 1, min_op_pos - 1));
+                            double y = double.Parse(Splitstring(formula, exist_op_div + 1, min_op_pos - 1));
                             ans = Super_Easy_calc(x, y, c);
                             return_value = ans + Splitstring(formula, min_op_pos, end);
                             return Calc(return_value);
@@ -214,7 +213,7 @@ namespace CALC
                         if (exist_plus_back == -1 && exist_minus_back == -1)
                         {
                             //後ろにも記号がなかった時
-                            int y = int.Parse(Splitstring(formula, exist_op_plus + 1, end));
+                            double y = double.Parse(Splitstring(formula, exist_op_plus + 1, end));
                             return_value = Super_Easy_calc(x, y, c);
                             return return_value;
                         }
@@ -222,7 +221,7 @@ namespace CALC
                         {
                             //後ろに記号があった時
                             int min_op_pos = Search_pos_min(exist_back) + exist_op_plus + 1;
-                            int y = int.Parse(Splitstring(formula, exist_op_plus + 1, min_op_pos - 1));
+                            double y = double.Parse(Splitstring(formula, exist_op_plus + 1, min_op_pos - 1));
                             ans = Super_Easy_calc(x, y, c);
                             return_value = ans + Splitstring(formula, min_op_pos, end);
                             return Calc(return_value);
@@ -230,13 +229,13 @@ namespace CALC
                     }
                     else {
                         //前に記号があった時
-                        int x = int.Parse(Splitstring(formula, exist_op_front + 1, exist_op_plus - 1));
+                        double x = double.Parse(Splitstring(formula, exist_op_front + 1, exist_op_plus - 1));
                         char c = formula[exist_op_plus];
                         if (exist_plus_back == -1 && exist_minus_back == -1)
                         {
                             //後ろに記号がなかった時
 
-                            int y = int.Parse(Splitstring(formula, exist_op_plus + 1, end));
+                            double y = double.Parse(Splitstring(formula, exist_op_plus + 1, end));
 
                             ans = Super_Easy_calc(x, y, c);
                             return_value = Splitstring(formula, start, exist_op_plus - 1) + ans;
@@ -246,7 +245,7 @@ namespace CALC
                         {
                             //後ろに記号があった時
                             int min_op_pos = Search_pos_min(exist_back) + exist_op_plus + 1;
-                            int y = int.Parse(Splitstring(formula, exist_op_plus + 1, min_op_pos - 1));
+                            double y = double.Parse(Splitstring(formula, exist_op_plus + 1, min_op_pos - 1));
                             ans = Super_Easy_calc(x, y, c);
                             return_value = Splitstring(formula, start, exist_op_plus - 1) + ans + Splitstring(formula, min_op_pos, end);
                             return Calc(return_value);
@@ -256,19 +255,19 @@ namespace CALC
                 int exist_op_minus = Search_op(formula, op[5], start, end);
                 if (exist_op_minus!=-1) {
                     //引き算記号があった時
-                    int x = int.Parse(Splitstring(formula, start, exist_op_minus - 1)); //前に記号はもう来ない
+                    double x = double.Parse(Splitstring(formula, start, exist_op_minus - 1)); //前に記号はもう来ない
                     char c = formula[exist_op_minus];
                     int exist_back = Search_op(formula,op[5],exist_op_minus+1,end);
                     if (exist_back!=-1) {
                         //後ろに記号がある
-                        int y = int.Parse(Splitstring(formula,exist_op_minus+1,exist_back-1));
+                        double y = double.Parse(Splitstring(formula,exist_op_minus+1,exist_back-1));
                         ans = Super_Easy_calc(x, y, c);
                         return_value = ans + Splitstring(formula, exist_op_minus + 1, end);
                         return Calc(return_value);
                     }
                     else {
                         //後ろに記号がない
-                        int y = int.Parse(Splitstring(formula, exist_op_minus + 1, end));
+                        double y = double.Parse(Splitstring(formula, exist_op_minus + 1, end));
                         return_value = Super_Easy_calc(x, y, c);
                         return return_value;
                     }
